fix ESlint automatic + refactor containers directory naming for consistency-convention + pass getRandomNickname into class
