+++ src/client/components/tetrisView/TetrisView.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Tetris from '../../containers/tetris/Tetris';
-import Players from '../../containers/players/Players';
+import Tetris from '../../containers/Tetris/Tetris';
+import Players from '../../containers/Players/Players';
 
 import './TetrisView.scss';
