+++ src/client/containers/Score/Score.scss
