+++ src/client/containers/SpectrumContainer/SpectrumContainer.js
