+++ src/client/middlewares/peerRTC.js
@@ -4,7 +4,7 @@ export default peer => ({ dispatch, getState }) => {
   peer.on('open', (id) => {
     console.log(`My peer ID is: ${id}`, getState());
     const { socket } = getState();
-    socket.emit('/game', { path: '/join', id, room: window.location.hash.slice(1), socket: socket.id });
+    socket.emit('/Game', { path: '/join', id, room: window.location.hash.slice(1), socket: socket.id });
   });
 
   if (peer) {
