+++ src/server/config/index.js
@@ -1,6 +1,6 @@
 export default {
   db: {
-    url: process.env.DB_URL || 'mongodb://localhost:27017/red-tetris',
+    url: process.env.DB_URL || 'mongodb://localhost:27017/red-Tetris',
   },
   port: process.env.PORT || 3000,
   env: process.env.NODE_ENV || 'development',
